requirements.txt
faust==1.10.4
kafka-python~=2.2.6
six~=1.17.0

pip~=25.1.1
attrs~=25.3.0
wheel~=0.43.0
cryptography~=44.0.2
mypy-extensions~=1.0.0
mode~=4.3.2
yarl~=1.20.0
colorlog~=6.9.0
pytz~=2024.2
idna~=3.6
multidict~=6.4.3
propcache~=0.3.1
opentracing~=1.3.0
click~=7.1.2
colorclass~=2.2.2
aiohttp-cors~=0.8.1
aiohttp~=3.11.18
croniter~=6.0.0
terminaltables~=3.1.10
venusian~=1.2.0
typing_extensions~=4.11.0
PyYAML~=6.0.1
aiosignal~=1.3.2
frozenlist~=1.6.0
aiohappyeyeballs~=2.6.1
pytest~=7.4.4
python-dateutil~=2.9.0.post0
setuptools~=69.5.1
packaging~=24.0
Pygments~=2.17.2
pandas~=2.2.3